backend/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
SQLAlchemy==2.0.40
pydantic==2.12.5
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
python-dotenv==1.2.1
slowapi==0.1.9
python-multipart==0.0.21
